[FEATURE] Add an aspect to allow fluid template to use dotted property path access
By default fluid will return a null object when he tries to access to a
translated property through this format "object.property.locale".
When the standard ObjectAccessorNode return NULL, we reiterate though property
path to see if there is a translatable object, in that case we try to find the
wanted property and the locale to get the translated value.

Resolves: #12